[#179282600] Implements common webview component with error and loading handlers
